Figured out 404 user-journey
Basically when a user hits 404, you really want them to hit a rendered not hydrated page. This is so if you use dynamic content, you defer to React Router DOM to generate the page on the client. This means dynamic content is reliably generated and only static pages (documented in routes) and pregenerated / hydrated. Otherwise, known pages are server-rendered and hydrated. The benefit of this approach is that you can mix-and-match client and server generation.